# server.js
const express = require('express');
const swaggerUI = require('swagger-ui-express');
const swaggerDocument = require('./openapi.json');
const app = express();
const PORT = 3000;

app.use(express.json());
app.use('/docs', swaggerUI.serve, swaggerUI.setup(swaggerDocument));

let tasks = [ 
  { id: 1, title: "Do homework", completed: false },
  { id: 2, title: "Walk the dog", completed: true },
  { id: 3, title: "Read a book", completed: false }
];

app.get('/', (req, res) => {
  res.json({
    name: "Task API",
    version : "1.0",
    endpoints:  ["/tasks"]
   });
});

app.get('/tasks', (req, res) => {
  res.json(tasks);
});

app.get('/tasks/:id', (req, res) => {
  const id = parseInt(req.params.id);
  const task = tasks.find(t => t.id === id);

  if (!task) {
    return res.status(404).json({ error: `Task ${id} not found` });
  }

  res.json(task);
});

app.post('/tasks', (req, res) => {
  const {title} = req.body;

  if (!title || title.trim() === "") {
    return res.status(400).json({ error: "Title is required" });
  }

  const newID = tasks.length > 0
    ? Math.max(...tasks.map(t => t.id)) + 1
    : 1;

  const newTask = {
    id: newID,
    title: title,
    done: false
  };

  tasks.push(newTask);
  res.status(201).json(newTask);
});

app.put('/tasks/:id', (req, res) => {
    const id = parseInt(req.params.id);
    const task= tasks.find(t => t.id === id);

    if (!task) {
      return res.status(404).json({ error: `Task ${id} not found` });
    }

    const { title, done } = req.body;

    if (title !== undefined && title.trim() === "") {
      return res.status(400).json({error: 'Title cannot be empty'})
    }

    if (title != undefined) task.title = title;
    if (done != undefined) task.done = done;

    res.json(task);
  });

  app.delete('/tasks/:id', (req,res) => {
    const id = parseInt(req.params.id);
    const index = tasks.findIndex(t => t.id === id);

    if(index === -1) {
      return res.status(400).json({error: `Task ${id} not found`})
    }

    tasks.splice(index, 1);
    res.status(204).send();

  });

app.get('/health', (req, res) => {
  res.json({ status: 'OK' });
});

app.get('/time', (req, res) => {
  res.json({ currentTime: new Date().toISOString() });
});

app.listen(PORT, () => {
  console.log(`Server running at http://localhost:${PORT}`);
});